+++ tanks-backend/TanksServer/Models/PixelPosition.cs
@@ -4,12 +4,12 @@ using TanksServer.GameLogic;
 namespace TanksServer.Models;
 
 [DebuggerDisplay("({X} | {Y})")]
-internal readonly struct PixelPosition(int x, int y)
+internal readonly struct PixelPosition(ulong x, ulong y)
 {
-    public int X { get; } = (x + MapService.PixelsPerRow) % MapService.PixelsPerRow;
-    public int Y { get; } = (y + MapService.PixelsPerColumn) % MapService.PixelsPerColumn;
+    public ulong X { get; } = (x + MapService.PixelsPerRow) % MapService.PixelsPerRow;
+    public ulong Y { get; } = (y + MapService.PixelsPerColumn) % MapService.PixelsPerColumn;
 
-    public void Deconstruct(out int x, out int y)
+    public void Deconstruct(out ulong x, out ulong y)
     {
         x = X;
         y = Y;
